requirements.txt
numpy>=1.24
Pillow>=10.0
matplotlib>=3.7
scipy>=1.10
